+++ packages/fz-design/components/message/src/interface.ts
@@ -1,1 +1,13 @@
+import type { ComponentType } from '../../../types'
+
 export type { MessageProps } from './props'
+
+export type MessageType = ComponentType
+
+export type MessagePlacement =
+  | 'top'
+  | 'top-left'
+  | 'top-right'
+  | 'bottom'
+  | 'bottom-left'
+  | 'bottom-right'