+++ packages/fz-design/types/component.ts
@@ -1,6 +1,10 @@
 import type { Component } from 'vue';
 
+/** 组件大小 - 尺寸 */
 export type ComponentSize = '' | 'small' | 'default' | 'large';
+
+/** 组件类型 - 颜色 */
+export type ComponentType = 'primary' | 'success' | 'info' | 'warning' | 'error' | 'default';
 
 /**
  * svg图标组件