+++ packages/fz-design/utils/lodash/lang.ts
@@ -3,54 +3,55 @@ export const isClient = typeof document !== 'undefined'
 export const isFirefox = (): boolean =>
   isClient && /firefox/i.test(window.navigator.userAgent)
 
-export const isBoolean = (val: any): boolean =>
-  Object.prototype.toString.call(val) === '[object Boolean]'
+/**
+ * 检测一个值是否为指定的类型
+ *
+ * @param { * } value 要检测的值
+ * @param { string } type 预期类型
+ * @returns { boolean } 这个值是否为传入的类型
+ */
+export const is = (value: unknown, type: string): boolean => {
+  return Object.prototype.toString.call(value) === `[object ${type}]`
+}
 
-export const isNumber = (val: any): boolean =>
-  Object.prototype.toString.call(val) === '[object Number]'
+export const isBoolean = (val: unknown): boolean => is(val, 'Boolean')
 
-export const isUndefined = (val: any): boolean =>
-  Object.prototype.toString.call(val) === '[object Undefined]'
+export const isNumber = (val: unknown): val is number => is(val, 'Number')
 
-export const isNull = (val: any): boolean => val === null
+export const isUndefined = (val: unknown): val is undefined => is(val, 'Undefined')
 
-export const isNil = (val: any): boolean => isUndefined(val) || isNull(val)
+export const isMap = (val: unknown): val is Map<unknown, unknown> => is(val, 'Map')
 
-export const hasOwn = (val: object, key: string | symbol): key is keyof typeof val =>
+export const isSet = (val: unknown): val is Set<unknown>  => is(val, 'Set')
+
+export const isDate = (val: unknown): val is Date => is(val, 'Date')
+
+export const isRegExp = (val: unknown): val is RegExp => is(val, 'RegExp')
+
+export const isString = (val: unknown): val is string => is(val, 'String')
+
+export const isSymbol = (val: unknown): val is Symbol => is(val, 'Symbol')
+
+export const isObject = (val: unknown): val is Record<string | number | symbol, unknown> => is(val, 'Object')
+
+export const isFunction = (val: unknown): boolean => typeof val === 'function'
+
+export const isNull = (val: unknown): boolean => val === null
+
+export const isNil = (val: unknown): val is any => isUndefined(val) || isNull(val)
+
+export const hasOwn = (val: object, key: string | symbol): boolean =>
   Object.prototype.hasOwnProperty.call(val, key)
 
 export const isArray = Array.isArray
 
-export const isMap = (val: unknown): val is Map<any, any> =>
-  Object.prototype.toString.call(val) === '[object Map]'
-
-export const isSet = (val: unknown): val is Set<any> =>
-  Object.prototype.toString.call(val) === '[object Set]'
-
-export const isDate = (val: unknown): val is Date =>
-  Object.prototype.toString.call(val) === '[object Date]'
-
-export const isRegExp = (val: unknown): val is RegExp =>
-  Object.prototype.toString.call(val) === '[object RegExp]'
-
-export const isFunction = (val: unknown): val is Function => typeof val === 'function'
-
-export const isString = (val: any): boolean =>
-  Object.prototype.toString.call(val) === '[object String]'
-export const isSymbol = (val: any): boolean =>
-  Object.prototype.toString.call(val) === '[object Symbol]'
-export const isObject = (val: unknown): val is Record<any, any> =>
-  val !== null && typeof val === 'object'
-
-export const isPromise = <T = any>(val: unknown): val is Promise<T> => {
+export const isPromise = <T = unknown>(val: unknown): val is Promise<T> => {
   return (
     (isObject(val) || isFunction(val)) &&
     isFunction((val as any).then) &&
     isFunction((val as any).catch)
   )
 }
-
-export const isEmpty = (val: any): boolean => isUndefined(val) || val === null
 
 export const isElement = (e: unknown): e is Element => {
   if (typeof Element === 'undefined') return false